requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
boto3==1.35.10
botocore==1.35.10
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework-simplejwt==5.3.1
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
flower==2.0.1
humanize==4.10.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
model-bakery==1.19.5
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
python-decouple==3.8
# use binary version if you have a trouble with installation.
# psycopg2-binary==2.9.9
psycopg==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
hiredis==3.0.0
referencing==0.35.1
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
